refactor(sos): trim orchestrator below 500 LOC; extract session-runner + venture-routing
Follow-up to sos/* split — orchestrator was still 524 LOC. Split out:
- sos/session-runner.ts (386 LOC): in-venture happy path —
  backfillPriorSessionActivity, fetchCiAlerts, fetchFleetHealth,
  fetchHandoffs, fetchFullModeData, fetchSessionContext, buildSosResult,
  runValidVentureSession.
- sos/venture-routing.ts (96 LOC): not-in-venture branches —
  handleVentureNavigation, handleVentureSelection.

sos.ts now 146 LOC: argument parsing, env/key validation, ventures fetch,
dispatch. Public API and test re-exports preserved.

46/46 sos tests pass.
